requirements.txt
numpy==1.19.2
scipy==1.4.1
matplotlib==3.1.2
tensorflow==2.4.1
opencv-python==4.5.1.48
annoy==1.17.0

Django==3.2.12
django-cors-headers==3.2.1
django-filter==2.4.0
PyJWT==1.7.1
django-graphql-jwt==0.3.0

graphene==2.1.8
graphene-django==2.8.0
Markdown==3.1.1
Pillow==9.0.1

inotify==0.2.10
asyncinotify==2.0.2
python-dateutil==2.8.1
pytz==2019.3
pyshp==2.1.0
requests==2.22.0

python-redis-lock==3.5.0
redis==3.3.11

psycopg2-binary==2.8.4
gunicorn==20.0.4

pytest==5.3.4
pytest-django==3.8.0
mock==3.0.5
factory-boy==2.12.0
coverage==5.0.3
colorlog==5.0.1
